requirements.txt
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
git+https://github.com/modelcontextprotocol/python-sdk.git@main#egg=mcp

# Development dependencies
pytest>=7.0.0
black>=22.1.0
flake8>=4.0.1
mypy>=0.931
isort>=5.10.1
pytest-cov>=3.0.0 